started to move events operations to repo
