+++ shipyard-controller/db/mongodb_event_repo_test.go
@@ -203,3 +203,76 @@ func GenerateTraceForRootEvent(projectName, stageName, serviceName string, rootE
 
 	return result
 }
+
+//func Test_GetEventsWithRetry(t *testing.T) {
+//	eventAvailable := false
+//	type fields struct {
+//		projectRepo db.ProjectMVRepo
+//		eventRepo   db.EventRepo
+//	}
+//	type args struct {
+//		project string
+//		filter  common.EventFilter
+//		status  common.EventStatus
+//	}
+//	tests := []struct {
+//		name    string
+//		fields  fields
+//		args    args
+//		want    []models.Event
+//		wantErr bool
+//	}{
+//		{
+//			name: "get event",
+//			fields: fields{
+//				projectRepo: nil,
+//				eventRepo: &db_mock.EventRepoMock{
+//					GetEventsWithRetryFunc: func(project string, filter common.EventFilter, status common.EventStatus, retries int) ([]models.Event, error) {
+//						return []models.Event{fake.GetTestTriggeredEvent()}, nil
+//					},
+//				},
+//			},
+//			args: args{
+//				project: "test-project",
+//				filter:  common.EventFilter{},
+//				status:  common.TriggeredEvent,
+//			},
+//			want:    []models.Event{fake.GetTestTriggeredEvent()},
+//			wantErr: false,
+//		},
+//		{
+//			name: "get event after retry",
+//			fields: fields{
+//				projectRepo: nil,
+//				eventRepo: &db_mock.EventRepoMock{
+//					GetEventsWithRetryFunc: func(project string, filter common.EventFilter, status common.EventStatus, retries int) ([]models.Event, error) {
+//						if eventAvailable {
+//							return []models.Event{fake.GetTestTriggeredEvent()}, nil
+//						}
+//						eventAvailable = true
+//						return nil, db.ErrNoEventFound
+//					},
+//				},
+//			},
+//			args: args{
+//				project: "test-project",
+//				filter:  common.EventFilter{},
+//				status:  common.TriggeredEvent,
+//			},
+//			want:    []models.Event{fake.GetTestTriggeredEvent()},
+//			wantErr: false,
+//		},
+//	}
+//	for _, tt := range tests {
+//		t.Run(tt.name, func(t *testing.T) {
+//			got, err := em.eventRepo.GetEventsWithRetry(tt.args.project, tt.args.filter, tt.args.status, 1)
+//			if (err != nil) != tt.wantErr {
+//				t.Errorf("getEvents() error = %v, wantErr %v", err, tt.wantErr)
+//				return
+//			}
+//			if !reflect.DeepEqual(got, tt.want) {
+//				t.Errorf("getEvents() got = %v, want %v", got, tt.want)
+//			}
+//		})
+//	}
+//}
