+++ shipyard-controller/db/repos.go
@@ -57,16 +57,16 @@ type EventQueueRepo interface {
 
 //go:generate moq --skip-ensure -pkg db_mock -out ./mock/eventrepo_mock.go . EventRepo
 type EventRepo interface {
-	// GetEvents gets all events of a project, based on the provided filter
 	GetEvents(project string, filter common.EventFilter, status ...common.EventStatus) ([]models.Event, error)
-	// GetRootEvents returns all root events of a project
 	GetRootEvents(params models.GetRootEventParams) (*models.GetEventsResult, error)
-	// InsertEvent inserts an event into the collection of the specified project
 	InsertEvent(project string, event models.Event, status common.EventStatus) error
-	// DeleteEvent deletes an event from the collection
 	DeleteEvent(project string, eventID string, status common.EventStatus) error
-	// DeleteEventCollections godoc
 	DeleteEventCollections(project string) error
+	GetStartedEventsForTriggeredID(eventScope *models.EventScope) ([]models.Event, error)
+	GetEventsWithRetry(project string, filter common.EventFilter, status common.EventStatus, nrRetries int) ([]models.Event, error)
+	GetTaskSequenceTriggeredEvent(eventScope models.EventScope, taskSequenceName string) (*models.Event, error)
+	DeleteAllFinishedEvents(eventScope models.EventScope) error
+	GetFinishedEvents(eventScope models.EventScope) ([]models.Event, error)
 }
 
 // ProjectRepo is an interface to access projects
